+++ lizard_portal/media/lizard_portal/lizard/portlet/WorkspacePortlet.js
@@ -74,6 +74,7 @@
     tools: [
       {
         type: 'empty',
+        tooltip: 'Workspace legen',
         handler: function(e, target, panelHeader, tool) {
           var portlet;
           portlet = panelHeader.ownerCt;
@@ -81,6 +82,7 @@
         }
       }, {
         type: 'save',
+        tooltip: 'Workspace opslaan',
         handler: function(e, target, panelHeader, tool) {
           var portlet;
           portlet = panelHeader.ownerCt;
@@ -101,6 +103,7 @@
         }
       }, {
         type: 'gear',
+        tooltip: 'Workspace beheer',
         handler: function(e, target, panelHeader, tool) {
           var a, form_window, portlet;
           portlet = panelHeader.ownerCt;
@@ -204,6 +207,7 @@
         }
       }, {
         type: 'delete',
+        tooltip: 'Workspace item verwijderen (na selectie)',
         handler: function(e, target, panelHeader, tool) {
           var portlet, records;
           portlet = panelHeader.ownerCt;